# Optimize getting config item values 