--- a/docs/changes/config/17461.perf.rst
+++ b/docs/changes/config/17461.perf.rst
@@ -1,0 +1,1 @@
+Optimize getting config item values.