@saimn 2024-11-26
Forgot to add some timings.
Before:
```
In [2]: %timeit fits.conf.enable_record_valued_keyword_cards
2.69 μs ± 502 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [3]: %timeit fits.getdata("astropy/io/fits/tests/data/tb.fits")
1.73 ms ± 338 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

After:
```
In [2]: %timeit fits.conf.enable_record_valued_keyword_cards
55.2 ns ± 0.163 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)

In [3]: %timeit fits.getdata("astropy/io/fits/tests/data/tb.fits")
1.09 ms ± 6.35 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```