@mhvk 2025-05-13
Unfortunately, this caching of config items introduced a bug, in that `reload_config` does not reset the values cached here - see #18138.